+++ tests/test_openapi_v3.py
@@ -1,3 +1,4 @@
+from blacksheep.server.routing import RoutesRegistry
 from dataclasses import dataclass
 from enum import IntEnum
 from typing import ForwardRef, Generic, List, Optional, TypeVar, Union
@@ -14,10 +15,23 @@ from blacksheep.server.openapi.exceptions import (
     UnsupportedUnionTypeException,
 )
 from blacksheep.server.openapi.v3 import OpenAPIHandler, check_union
-from openapidocs.common import Format
+from openapidocs.common import Format, Serializer
 from openapidocs.v3 import Info, Reference, Schema, ValueType
 
 T = TypeVar("T")
+U = TypeVar("U")
+
+
+@dataclass
+class Cat:
+    id: int
+    name: str
+
+
+@dataclass
+class Combo(Generic[T, U]):
+    item_one: T
+    item_two: U
 
 
 @dataclass
@@ -56,24 +70,36 @@ class Ufo:
     c: str
 
 
+@pytest.fixture
+def app() -> Application:
+    app = Application()
+    app.controllers_router = RoutesRegistry()
+    return app
+
+
+@pytest.fixture
+def docs() -> OpenAPIHandler:
+    # example documentation
+    return OpenAPIHandler(info=Info("Example", "0.0.1"))
+
+
+@pytest.fixture
+def serializer() -> Serializer:
+    return Serializer()
+
+
 @pytest.mark.asyncio
-async def test_raises_for_started_app():
+async def test_raises_for_started_app(docs):
     app = Application()
 
     await app.start()
-
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
 
     with pytest.raises(TypeError):
         docs.bind_app(app)
 
 
 @pytest.mark.asyncio
-async def test_raises_for_duplicated_content_example():
-    app = Application()
-
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+async def test_raises_for_duplicated_content_example(docs, app):
     @app.router.get("/")
     @docs(
         responses={
@@ -89,8 +115,7 @@ async def test_raises_for_duplicated_content_example():
 
 
 @pytest.mark.asyncio
-def test_raises_for_union_type():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
+def test_raises_for_union_type(docs):
     with pytest.raises(UnsupportedUnionTypeException):
         docs.get_schema_by_type(Union[Foo, Ufo])
 
@@ -108,9 +133,7 @@ def test_check_union(annotation, expected_result):
     assert check_union(annotation) == tuple(expected_result)
 
 
-def test_register_schema_can_handle_classes_with_same_name():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+def test_register_schema_can_handle_classes_with_same_name(docs):
     @dataclass
     class FooX:
         x: str
@@ -133,9 +156,7 @@ def test_register_schema_can_handle_classes_with_same_name():
     assert "a" in foo_schema.properties
 
 
-def test_register_schema_handles_repeated_calls():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+def test_register_schema_handles_repeated_calls(docs):
     docs.register_schema_for_type(Foo)
     docs.register_schema_for_type(Foo)
     docs.register_schema_for_type(Foo)
@@ -150,9 +171,7 @@ def test_register_schema_handles_repeated_calls():
     assert "a" in foo_schema.properties
 
 
-def test_handles_forward_refs():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+def test_handles_forward_refs(docs):
     @dataclass
     class Friend:
         foo: "Foo"
@@ -169,9 +188,7 @@ def test_handles_forward_refs():
     assert friend_schema.properties["foo"] == Reference(ref="#/components/schemas/Foo")
 
 
-def test_register_schema_for_enum():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+def test_register_schema_for_enum(docs):
     docs.register_schema_for_type(FooLevel)
 
     assert docs.components.schemas is not None
@@ -185,25 +202,20 @@ def test_register_schema_for_enum():
     assert schema.enum == [x.value for x in FooLevel]
 
 
-def test_try_get_schema_for_enum_returns_none_for_not_enum():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+def test_try_get_schema_for_enum_returns_none_for_not_enum(docs):
     assert docs._try_get_schema_for_enum(Foo) is None
 
 
-def test_get_parameters_returns_non_for_object_without_binders():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
+def test_get_parameters_returns_non_for_object_without_binders(docs):
     assert docs.get_parameters(Foo) is None
     assert docs.get_request_body(Foo) is None
 
 
-def test_get_content_from_response_info_returns_none_for_missing_content():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
+def test_get_content_from_response_info_returns_none_for_missing_content(docs):
     assert docs._get_content_from_response_info(None) is None
 
 
-def test_get_schema_by_type_returns_reference_for_forward_ref():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
+def test_get_schema_by_type_returns_reference_for_forward_ref(docs):
     assert docs._get_schema_by_type(ForwardRef("Foo")) == Reference(
         "#/components/schemas/Foo"
     )
@@ -230,17 +242,16 @@ def test_get_spec_path_preferred_format(
     assert docs.get_spec_path() == expected_result
 
 
-def test_get_spec_path_raises_for_unsupported_preferred_format():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
+def test_get_spec_path_raises_for_unsupported_preferred_format(docs):
     docs.preferred_format = "NOPE"  # type: ignore
 
     with pytest.raises(OpenAPIEndpointException):
         docs.get_spec_path()
 
 
-def test_register_schema_for_generic_with_list():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+def test_register_schema_for_generic_with_list(
+    docs: OpenAPIHandler, serializer: Serializer
+):
     docs.register_schema_for_type(PaginatedSet[Foo])
 
     assert docs.components.schemas is not None
@@ -248,12 +259,59 @@ def test_register_schema_for_generic_with_list():
 
     assert isinstance(schema, Schema)
 
-    assert schema is not None
+    yaml = serializer.to_yaml(docs.generate_documentation(Application()))
+
+    assert (
+        yaml.strip()
+        == """
+openapi: 3.0.3
+info:
+    title: Example
+    version: 0.0.1
+paths: {}
+components:
+    schemas:
+        Foo:
+            type: object
+            required:
+            - a
+            - b
+            properties:
+                a:
+                    type: string
+                    nullable: false
+                b:
+                    type: boolean
+                    nullable: false
+                level:
+                    type: integer
+                    nullable: false
+                    enum:
+                    - 1
+                    - 2
+                    - 3
+        PaginatedSet<Foo>:
+            type: object
+            required:
+            - items
+            - total
+            properties:
+                items:
+                    type: array
+                    nullable: false
+                    items:
+                        $ref: '#/components/schemas/Foo'
+                total:
+                    type: integer
+                    format: int64
+                    nullable: false
+""".strip()
+    )
 
 
-def test_register_schema_for_multiple_generic_with_list():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+def test_register_schema_for_multiple_generic_with_list(
+    docs: OpenAPIHandler, serializer: Serializer
+):
     docs.register_schema_for_type(PaginatedSet[Foo])
     docs.register_schema_for_type(PaginatedSet[Ufo])
 
@@ -264,12 +322,86 @@ def test_register_schema_for_multiple_generic_with_list():
     schema = docs.components.schemas["PaginatedSet<Ufo>"]
     assert isinstance(schema, Schema)
 
-    assert schema is not None
+    yaml = serializer.to_yaml(docs.generate_documentation(Application()))
+
+    assert (
+        yaml.strip()
+        == """
+openapi: 3.0.3
+info:
+    title: Example
+    version: 0.0.1
+paths: {}
+components:
+    schemas:
+        Foo:
+            type: object
+            required:
+            - a
+            - b
+            properties:
+                a:
+                    type: string
+                    nullable: false
+                b:
+                    type: boolean
+                    nullable: false
+                level:
+                    type: integer
+                    nullable: false
+                    enum:
+                    - 1
+                    - 2
+                    - 3
+        PaginatedSet<Foo>:
+            type: object
+            required:
+            - items
+            - total
+            properties:
+                items:
+                    type: array
+                    nullable: false
+                    items:
+                        $ref: '#/components/schemas/Foo'
+                total:
+                    type: integer
+                    format: int64
+                    nullable: false
+        Ufo:
+            type: object
+            required:
+            - b
+            - c
+            properties:
+                b:
+                    type: boolean
+                    nullable: false
+                c:
+                    type: string
+                    nullable: false
+        PaginatedSet<Ufo>:
+            type: object
+            required:
+            - items
+            - total
+            properties:
+                items:
+                    type: array
+                    nullable: false
+                    items:
+                        $ref: '#/components/schemas/Ufo'
+                total:
+                    type: integer
+                    format: int64
+                    nullable: false
+""".strip()
+    )
 
 
-def test_register_schema_for_generic_with_property():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+def test_register_schema_for_generic_with_property(
+    docs: OpenAPIHandler, serializer: Serializer
+):
     docs.register_schema_for_type(Validated[Foo])
 
     assert docs.components.schemas is not None
@@ -277,12 +409,55 @@ def test_register_schema_for_generic_with_property():
 
     assert isinstance(schema, Schema)
 
-    assert schema is not None
+    yaml = serializer.to_yaml(docs.generate_documentation(Application()))
+
+    assert (
+        yaml.strip()
+        == """
+openapi: 3.0.3
+info:
+    title: Example
+    version: 0.0.1
+paths: {}
+components:
+    schemas:
+        Foo:
+            type: object
+            required:
+            - a
+            - b
+            properties:
+                a:
+                    type: string
+                    nullable: false
+                b:
+                    type: boolean
+                    nullable: false
+                level:
+                    type: integer
+                    nullable: false
+                    enum:
+                    - 1
+                    - 2
+                    - 3
+        Validated<Foo>:
+            type: object
+            required:
+            - data
+            - error
+            properties:
+                data:
+                    $ref: '#/components/schemas/Foo'
+                error:
+                    type: string
+                    nullable: false
+""".strip()
+    )
 
 
-def test_register_schema_for_generic_sub_property():
-    docs = OpenAPIHandler(info=Info("Example", "0.0.1"))
-
+def test_register_schema_for_generic_sub_property(
+    docs: OpenAPIHandler, serializer: Serializer
+):
     docs.register_schema_for_type(Validated[Foo])
     docs.register_schema_for_type(SubValidated[Foo])
 
@@ -291,4 +466,133 @@ def test_register_schema_for_generic_sub_property():
 
     assert isinstance(schema, Schema)
 
-    assert schema is not None
+    yaml = serializer.to_yaml(docs.generate_documentation(Application()))
+
+    assert (
+        yaml.strip()
+        == """
+openapi: 3.0.3
+info:
+    title: Example
+    version: 0.0.1
+paths: {}
+components:
+    schemas:
+        Foo:
+            type: object
+            required:
+            - a
+            - b
+            properties:
+                a:
+                    type: string
+                    nullable: false
+                b:
+                    type: boolean
+                    nullable: false
+                level:
+                    type: integer
+                    nullable: false
+                    enum:
+                    - 1
+                    - 2
+                    - 3
+        Validated<Foo>:
+            type: object
+            required:
+            - data
+            - error
+            properties:
+                data:
+                    $ref: '#/components/schemas/Foo'
+                error:
+                    type: string
+                    nullable: false
+        SubValidated<Foo>:
+            type: object
+            required:
+            - sub
+            properties:
+                sub:
+                    $ref: '#/components/schemas/Validated<Foo>'
+""".strip()
+    )
+
+
+@pytest.mark.asyncio
+async def test_register_schema_for_multi_generic(
+    app: Application, docs: OpenAPIHandler, serializer: Serializer
+):
+    @app.route("/combo")
+    def combo_example() -> Combo[Cat, Foo]:
+        ...
+
+    docs.bind_app(app)
+    await app.start()
+
+    yaml = serializer.to_yaml(docs.generate_documentation(app))
+
+    assert (
+        yaml.strip()
+        == """
+openapi: 3.0.3
+info:
+    title: Example
+    version: 0.0.1
+paths:
+    /combo:
+        get:
+            responses:
+                '200':
+                    description: Success response
+                    content:
+                        application/json:
+                            schema:
+                                $ref: '#/components/schemas/Combo<Cat, Foo>'
+            operationId: combo_example
+components:
+    schemas:
+        Cat:
+            type: object
+            required:
+            - id
+            - name
+            properties:
+                id:
+                    type: integer
+                    format: int64
+                    nullable: false
+                name:
+                    type: string
+                    nullable: false
+        Foo:
+            type: object
+            required:
+            - a
+            - b
+            properties:
+                a:
+                    type: string
+                    nullable: false
+                b:
+                    type: boolean
+                    nullable: false
+                level:
+                    type: integer
+                    nullable: false
+                    enum:
+                    - 1
+                    - 2
+                    - 3
+        Combo<Cat, Foo>:
+            type: object
+            required:
+            - item_one
+            - item_two
+            properties:
+                item_one:
+                    $ref: '#/components/schemas/Cat'
+                item_two:
+                    $ref: '#/components/schemas/Foo'
+""".strip()
+    )